% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seifa.R
\docType{data}
\name{seifa_exclusions}
\alias{seifa_exclusions}
\title{SEIFA Excluded Areas}
\format{a data frame / `data.table` with 7 columns
\describe{
\item{\code{SA1_7DIGITCODE_2016}}{A 7 digit identifier for the Statistical Area 1 (SA1)}
\item{\code{SA1_11DIGITCODE_2016}}{A 11 digit identifier for the Statistical Area 1 (SA1)}
\item{\code{usual_resident_population}}{Number of people usually resident in the SA1}
\item{\code{No_index_IRSAD}}{Indicates exclusion from the Index of Relative Socio-Economic Advantage and Disadvantage}
\item{\code{No_index_IRSD}}{Indicates exclusion from the Index of Relative Socio-Economic Disadvantage}
\item{\code{No_index_IER}}{Indicates exclusion from the Index of Economic Resources}
\item{\code{No_index_IEO}}{Indicates exclusion from the Index of Education and Occupation}
}}
\usage{
seifa_exclusions
}
\description{
Census of Population and Housing: Socio-Economic Indexes for Areas (SEIFA),
Australia, 2016
}
\details{
2016 SA1 geographies that have been excluded from the indexes.

More information about this dataset can be found at the Australian Bureau of
Statistics website \href{http://www.abs.gov.au/ausstats/abs@.nsf/mf/2033.0.55.001}{SEIFA}
}
\keyword{datasets}
