% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seifa.R
\docType{data}
\name{seifa_IRSAD}
\alias{seifa_IRSAD}
\title{SEIFA Index of Relative Socio-Economic Advantage and Disadvantage (IRSAD)}
\format{a data frame / `data.table` with 11 columns
\describe{
\item{\code{SA1_7DIGITCODE_2016}}{A 7 digit identifier for the Statistical Area 1 (SA1)}
\item{\code{SA1_11DIGITCODE_2016}}{A 11 digit identifier for the Statistical Area 1 (SA1)}
\item{\code{usual_resident_population}}{Number of people usually resident in the SA1}
\item{\code{score}}{Index of Relative Socio-Economic Advantage and Disadvantage - numeric score}
\item{\code{ranking_in_aust_rank}}{Ranking 1:n across SA1 boundaries in Australia}
\item{\code{ranking_in_aust_decile}}{Rank decile across Australia}
\item{\code{ranking_in_aust_percentile}}{Rank percentile across Australia}
\item{\code{ranking_in_state_state}}{State of ranking}
\item{\code{ranking_in_state_rank}}{Ranking 1:n across SA1 boundaries in a state}
\item{\code{ranking_in_state_decile}}{Rank decile across state}
\item{\code{ranking_in_state_percentile}}{Rank percentile across state}
}}
\usage{
seifa_IRSAD
}
\description{
Census of Population and Housing: Socio-Economic Indexes for Areas (SEIFA),
Australia, 2016
}
\details{
2016 IRSAD data at Statistical Area 1 (SA1)

More information about this dataset can be found at the Australian Bureau of
Statistics website \href{http://www.abs.gov.au/ausstats/abs@.nsf/mf/2033.0.55.001}{SEIFA}
}
\keyword{datasets}
