% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seifa.R
\docType{data}
\name{seifa_summary}
\alias{seifa_summary}
\title{SEIFA Summary}
\format{a data frame / `data.table` with 11 columns
\describe{
\item{\code{SA1_7DIGITCODE_2016}}{A 7 digit identifier for the Statistical Area 1 (SA1)}
\item{\code{SA1_11DIGITCODE_2016}}{A 11 digit identifier for the Statistical Area 1 (SA1)}
\item{\code{IRSD_score}}{Index of Relative Socio-Economic Disadvantage - numeric score}
\item{\code{IRSD_decile}}{Index of Relative Socio-Economic Disadvantage - decile (across Australia)}
\item{\code{IRSAD_score}}{Index of Relative Socio-Economic Advantage and Disadvantage - numeric score}
\item{\code{IRSAD_decile}}{Index of Relative Socio-Economic Advantage and Disadvantage - decile (across Australia)}
\item{\code{IER_score}}{Index of Economic Resources - numeric score}
\item{\code{IER_decile}}{Index of Economic Resources - decile (across Australia)}
\item{\code{IEO_score}}{Index of Education and Occupation - numeric score}
\item{\code{IEO_decile}}{Index of Education and Occupation - decile (across Australia)}
}}
\usage{
seifa_summary
}
\description{
Census of Population and Housing: Socio-Economic Indexes for Areas (SEIFA),
Australia, 2016
}
\details{
2016 SEIFA summary data at Statistical Area 1 (SA1)

More information about this dataset can be found at the Australian Bureau of
Statistics website \href{http://www.abs.gov.au/ausstats/abs@.nsf/mf/2033.0.55.001}{SEIFA}
}
\keyword{datasets}
